+++ src/index.html
@@ -17,7 +17,7 @@
     </div>
 
     <div class="list-add">
-      <input type="text" placeholder="Add to your list...">
+      <input type="text" placeholder="Add to your list..." id="input-add">
       <button type="button" id="button-add">
         <img src="https://media.flaticon.com/dist/min/img/uicons/icons/rounded/regular/fi-rr-add.svg" width="14" alt="Add task">
       </button>
